added checking for vpc existence, changed structure
