+++ script/features/04_aws_structure_creation.feature
@@ -9,11 +9,13 @@ Feature: Creating AWS structure for usage with MongoDB
 		And terraform bucket name for AWS is terraform-state
 		And AWS region for AWS is eu-west-2
 		And terraform region for S3 is us-west-2
-		And namespace for AWS is Testing
+		And namespace for AWS is cluster_automation_test
 		And project name for AWS is aws-vpc
 		And public key path for AWS is ~/.ssh/id_rsa.pub
 		And terraform backend is set to S3 bucket
+		And VPC with name cluster_automation_test is doesn't exist
 		When terraform config for AWS is initialized
 		And terraform config for AWS is planned
 		And terraform config for AWS is applied
+		Then VPC with name cluster_automation_test is exists
