+++ script/features/steps/terraform_common.py
@@ -1,7 +1,7 @@
 from lettuce import step, world
 
 from test_config import TERRAFORM_DIR, RUNNER_TEMPLATE
-from utils import TFRunner, check_if_bucket_exists
+from utils import TFRunner
 
 world.tf_runners = RUNNER_TEMPLATE.copy()
 
@@ -24,14 +24,3 @@ def run_terraform_stage(step_instance, namespace, stage):
 		assert False, "Such operation doesn't exist"
 	
 	assert err == "", "Terraform config exited with error: {}".format(err)
-
-
-@step("S3 bucket called ([A-Za-z0-9\-_]*) is (created|deleted)")
-def check_bucket_state(step_instance, full_bucket_name, state):
-	exists = check_if_bucket_exists(full_bucket_name)
-	if state == "created":
-		assert exists, "The bucket called {} is not created".format(full_bucket_name)
-	elif state == "deleted":
-		assert not exists, "The bucket called {} is not deleted".format(full_bucket_name)
-	else:
-		assert False, "Such state doesn't exist"
